# per-world contact