--- a/mujoco_warp/_src/collision_sdf.py
+++ b/mujoco_warp/_src/collision_sdf.py
@@ -330,17 +330,16 @@ def _sdf_narrowphase(
   sdf_iterations: int,
   # Data out:
   ncon_out: wp.array(dtype=int),
-  contact_dist_out: wp.array(dtype=float),
-  contact_pos_out: wp.array(dtype=wp.vec3),
-  contact_frame_out: wp.array(dtype=wp.mat33),
-  contact_includemargin_out: wp.array(dtype=float),
-  contact_friction_out: wp.array(dtype=vec5),
-  contact_solref_out: wp.array(dtype=wp.vec2),
-  contact_solreffriction_out: wp.array(dtype=wp.vec2),
-  contact_solimp_out: wp.array(dtype=vec5),
-  contact_dim_out: wp.array(dtype=int),
-  contact_geom_out: wp.array(dtype=wp.vec2i),
-  contact_worldid_out: wp.array(dtype=int),
+  contact_dist_out: wp.array2d(dtype=float),
+  contact_pos_out: wp.array2d(dtype=wp.vec3),
+  contact_frame_out: wp.array2d(dtype=wp.mat33),
+  contact_includemargin_out: wp.array2d(dtype=float),
+  contact_friction_out: wp.array2d(dtype=vec5),
+  contact_solref_out: wp.array2d(dtype=wp.vec2),
+  contact_solreffriction_out: wp.array2d(dtype=wp.vec2),
+  contact_solimp_out: wp.array2d(dtype=vec5),
+  contact_dim_out: wp.array2d(dtype=int),
+  contact_geom_out: wp.array2d(dtype=wp.vec2i),
 ):
   tid = wp.tid()
 
@@ -518,7 +517,6 @@ def _sdf_narrowphase(
       contact_solimp_out,
       contact_dim_out,
       contact_geom_out,
-      contact_worldid_out,
     )
 
 
@@ -526,7 +524,7 @@ def _sdf_narrowphase(
 def sdf_narrowphase(m: Model, d: Data):
   wp.launch(
     _sdf_narrowphase,
-    dim=d.nconmax,
+    dim=d.nworld * d.nconmax,
     inputs=[
       m.geom_type,
       m.geom_condim,
@@ -594,6 +592,5 @@ def sdf_narrowphase(m: Model, d: Data):
       d.contact.solimp,
       d.contact.dim,
       d.contact.geom,
-      d.contact.worldid,
     ],
   )