inspired by the performance improvements from changing constraints from all-worlds to per-world #361, this pr investigates changing contacts from all-worlds to per-world.

tldr
performance seems to be about the same, but there might be other benefits from changing contacts to per-world.

**humanoid**

main
```
mjwarp-testspeed --function=step --mjcf=test_data/humanoid/humanoid.xml --batch_size=8192 --nconmax=196608 --njmax=64
```
```
Summary for 8192 parallel rollouts

 Total JIT time: 0.25 s
 Total simulation time: 16.96 s
 Total steps per second: 483,045
 Total realtime factor: 2,415.22 x
 Total time per step: 2070.20 ns
```

this pr
```
mjwarp-testspeed --function=step --mjcf=test_data/humanoid/humanoid.xml --batch_size=8192 --nconmax=24 --njmax=64
```
```
Summary for 8192 parallel rollouts

 Total JIT time: 0.88 s
 Total simulation time: 16.88 s
 Total steps per second: 485,250
 Total realtime factor: 2,426.25 x
 Total time per step: 2060.79 ns
```

**aloha + pot**

main
```
mjwarp-testspeed --function=step --mjcf=benchmark/aloha_pot/scene.xml --batch_size=8192 --nconmax=131136 --njmax=96 --cone=pyramidal --keyframe=0
```
```
Summary for 8192 parallel rollouts

 Total JIT time: 0.65 s
 Total simulation time: 33.28 s
 Total steps per second: 246,178
 Total realtime factor: 492.36 x
 Total time per step: 4062.09 ns
```

this pr
```
mjwarp-testspeed --function=step --mjcf=benchmark/aloha_pot/scene.xml --batch_size=8192 --nconmax=16 --njmax=96 --cone=pyramidal --keyframe=0
```
```
Summary for 8192 parallel rollouts

 Total JIT time: 0.58 s
 Total simulation time: 33.28 s
 Total steps per second: 246,161
 Total realtime factor: 492.32 x
 Total time per step: 4062.38 ns
```

additional benefits from changing contacts to per-world:

- `njmax` and `nconmax` would share per-world semantics
- i think `njmax` and `nconmax` could be moved from `Data` to `Model` and utilize the existing fields in `mjModel`

there might be performance improvements for elliptic friction cones. once open issues for elliptic friction cones are resolved we can revisit potential performance improvements from this pr.