@erikfrey 2025-07-28
OK interesting!  As mentioned in #361 I'm actually not sure we want per-world grouping as it challenges the design around heterogenous batches of simulation, and that we might go back at some point if we can figure out how to do it without sacrificing performance.

So let's keep this in our back pocket if we need it but I wouldn't be in a hurry to merge this without evidence of big speedups.

Separately we can adjust our API functions, e.g. we can make `make_data` and `put_data` accept per-world nconmax if it's more intuitive for the two params to be more aligned that way.